# Avoid the enumerator allocation in Dictionary's constructor