Fast path to avoid the enumerator allocation and overhead when the passed-in `dictionary` is `Dictionary<TKey, TValue>`.