@justinvp 2015-09-30
@jkotas, I updated the PR with a new commit that checks for `Dictionary<TKey, TValue>`  alone (no subclasses), to avoid the subtle behavior change for subclasses of `Dictionary<TKey, TValue>` that may have overridden the enumerator behavior. Some micro-benchmark numbers are below. What do you think?

Instead of using `Dictionary<TKey, TValue>`'s struct enumerator in the fast path, we can avoid the enumerator overhead by looping over the entries directly (along the lines of other dictionary methods like [`CopyTo`](https://github.com/dotnet/coreclr/blob/1572a8ab9b4f30459b3ae071cab2ae58e09e213d/src/mscorlib/src/System/Collections/Generic/Dictionary.cs#L256-L262)).

This is a little more code than the `foreach`, but it is faster, and from a cursory look at the full framework reference source, [most](http://referencesource.microsoft.com/#mscorlib/system/collections/generic/dictionary.cs,6009e5bfe5f70341,references) (perhaps all?) [callers](http://referencesource.microsoft.com/#mscorlib/system/collections/generic/dictionary.cs,c59275a169cb495a,references) of these constructor overloads are passing-in `Dictionary<TKey, TValue>`.

Below are some numbers from a 1,000,000 iteration micro-benchmark, run against a private release build of CoreCLR with crossgen'd mscorlib.dll on Windows 10 (64 bit). Source is available [here](https://gist.github.com/justinvp/fd4a78259f54e9516994).

I submitted a PR to CoreFX (https://github.com/dotnet/corefx/pull/3545) that adds test coverage of these constructor overloads. The tests pass when run locally before and after the changes.

| Type                        | Size | Before (ms) | After (ms) | Faster |
|-----------------------------|-----:|------------:|-----------:|-------:|
| `Dictionary<Object,Object>` |    0 |          80 |         39 |   106% |
| `Dictionary<Object,Object>` |    1 |         155 |         98 |    58% |
| `Dictionary<Object,Object>` |    2 |         192 |        109 |    76% |
| `Dictionary<Object,Object>` |    3 |         254 |        130 |    96% |
| `Dictionary<Object,Object>` |    4 |         308 |        166 |    85% |
| `Dictionary<Object,Object>` |    5 |         391 |        190 |   106% |
| `Dictionary<Object,Object>` |   10 |         631 |        352 |    79% |
| `Dictionary<Object,Object>` |   25 |        1368 |        817 |    67% |
| `Dictionary<Object,Object>` |   50 |        2733 |       1502 |    82% |
| `Dictionary<Object,Object>` |  100 |        5764 |       3031 |    90% |
| `Dictionary<String,String>` |    0 |          78 |         45 |    73% |
| `Dictionary<String,String>` |    1 |         165 |        107 |    54% |
| `Dictionary<String,String>` |    2 |         218 |        140 |    56% |
| `Dictionary<String,String>` |    3 |         297 |        170 |    75% |
| `Dictionary<String,String>` |    4 |         393 |        222 |    77% |
| `Dictionary<String,String>` |    5 |         422 |        246 |    71% |
| `Dictionary<String,String>` |   10 |         777 |        497 |    56% |
| `Dictionary<String,String>` |   25 |        1759 |       1159 |    52% |
| `Dictionary<String,String>` |   50 |        3498 |       2233 |    57% |
| `Dictionary<String,String>` |  100 |        6565 |       4476 |    47% |
| `Dictionary<Int64,Int64>`   |    0 |          69 |         36 |    93% |
| `Dictionary<Int64,Int64>`   |    1 |         122 |         75 |    63% |
| `Dictionary<Int64,Int64>`   |    2 |         153 |         87 |    76% |
| `Dictionary<Int64,Int64>`   |    3 |         182 |        102 |    78% |
| `Dictionary<Int64,Int64>`   |    4 |         214 |        126 |    70% |
| `Dictionary<Int64,Int64>`   |    5 |         245 |        147 |    66% |
| `Dictionary<Int64,Int64>`   |   10 |         398 |        241 |    65% |
| `Dictionary<Int64,Int64>`   |   25 |         861 |        509 |    69% |
| `Dictionary<Int64,Int64>`   |   50 |        1665 |        954 |    75% |
| `Dictionary<Int64,Int64>`   |  100 |        3168 |       1812 |    75% |
| `Dictionary<Int32,Int32>`   |    0 |          63 |         37 |    68% |
| `Dictionary<Int32,Int32>`   |    1 |         121 |         81 |    51% |
| `Dictionary<Int32,Int32>`   |    2 |         150 |         88 |    70% |
| `Dictionary<Int32,Int32>`   |    3 |         168 |        102 |    65% |
| `Dictionary<Int32,Int32>`   |    4 |         210 |        122 |    72% |
| `Dictionary<Int32,Int32>`   |    5 |         232 |        140 |    65% |
| `Dictionary<Int32,Int32>`   |   10 |         389 |        233 |    67% |
| `Dictionary<Int32,Int32>`   |   25 |         805 |        486 |    66% |
| `Dictionary<Int32,Int32>`   |   50 |        1535 |        913 |    68% |
| `Dictionary<Int32,Int32>`   |  100 |        2928 |       1723 |    70% |


| Type                                | Size | Before (ms) | After (ms) | Faster |
|-------------------------------------|-----:|------------:|-----------:|-------:|
| `DictionarySubclass<Object,Object>` |    0 |          84 |         82 |     2% |
| `DictionarySubclass<Object,Object>` |    1 |         154 |        160 |    -4% |
| `DictionarySubclass<Object,Object>` |    2 |         192 |        208 |    -7% |
| `DictionarySubclass<Object,Object>` |    3 |         235 |        247 |    -5% |
| `DictionarySubclass<Object,Object>` |    4 |         306 |        314 |    -3% |
| `DictionarySubclass<Object,Object>` |    5 |         344 |        354 |    -3% |
| `DictionarySubclass<Object,Object>` |   10 |         627 |        609 |     3% |
| `DictionarySubclass<Object,Object>` |   25 |        1376 |       1370 |     0% |
| `DictionarySubclass<Object,Object>` |   50 |        2691 |       2607 |     3% |
| `DictionarySubclass<Object,Object>` |  100 |        5401 |       5210 |     4% |
| `DictionarySubclass<String,String>` |    0 |          78 |         83 |    -7% |
| `DictionarySubclass<String,String>` |    1 |         169 |        173 |    -2% |
| `DictionarySubclass<String,String>` |    2 |         231 |        240 |    -4% |
| `DictionarySubclass<String,String>` |    3 |         293 |        278 |     5% |
| `DictionarySubclass<String,String>` |    4 |         391 |        400 |    -2% |
| `DictionarySubclass<String,String>` |    5 |         439 |        439 |     0% |
| `DictionarySubclass<String,String>` |   10 |         795 |        757 |     5% |
| `DictionarySubclass<String,String>` |   25 |        1756 |       1673 |     5% |
| `DictionarySubclass<String,String>` |   50 |        3514 |       3377 |     4% |
| `DictionarySubclass<String,String>` |  100 |        6555 |       7169 |    -9% |
| `DictionarySubclass<Int64,Int64>`   |    0 |          72 |         76 |    -6% |
| `DictionarySubclass<Int64,Int64>`   |    1 |         128 |        139 |    -8% |
| `DictionarySubclass<Int64,Int64>`   |    2 |         160 |        155 |     3% |
| `DictionarySubclass<Int64,Int64>`   |    3 |         190 |        184 |     4% |
| `DictionarySubclass<Int64,Int64>`   |    4 |         224 |        224 |     0% |
| `DictionarySubclass<Int64,Int64>`   |    5 |         254 |        262 |    -3% |
| `DictionarySubclass<Int64,Int64>`   |   10 |         407 |        400 |     2% |
| `DictionarySubclass<Int64,Int64>`   |   25 |         866 |        863 |     0% |
| `DictionarySubclass<Int64,Int64>`   |   50 |        1651 |       1653 |     0% |
| `DictionarySubclass<Int64,Int64>`   |  100 |        3188 |       3198 |     0% |
| `DictionarySubclass<Int32,Int32>`   |    0 |          77 |         70 |     9% |
| `DictionarySubclass<Int32,Int32>`   |    1 |         121 |        124 |    -3% |
| `DictionarySubclass<Int32,Int32>`   |    2 |         155 |        149 |     4% |
| `DictionarySubclass<Int32,Int32>`   |    3 |         170 |        182 |    -7% |
| `DictionarySubclass<Int32,Int32>`   |    4 |         207 |        210 |    -1% |
| `DictionarySubclass<Int32,Int32>`   |    5 |         230 |        234 |    -2% |
| `DictionarySubclass<Int32,Int32>`   |   10 |         378 |        387 |    -2% |
| `DictionarySubclass<Int32,Int32>`   |   25 |         814 |        828 |    -2% |
| `DictionarySubclass<Int32,Int32>`   |   50 |        1536 |       1586 |    -3% |
| `DictionarySubclass<Int32,Int32>`   |  100 |        2943 |       3031 |    -3% |


| Type                                | Size | Before (ms) | After (ms) | Faster |
|-------------------------------------|-----:|------------:|-----------:|-------:|
| `ReadOnlyDictionary<Object,Object>` |    0 |          85 |         89 |    -5% |
| `ReadOnlyDictionary<Object,Object>` |    1 |         158 |        172 |    -8% |
| `ReadOnlyDictionary<Object,Object>` |    2 |         215 |        217 |    -1% |
| `ReadOnlyDictionary<Object,Object>` |    3 |         243 |        244 |     0% |
| `ReadOnlyDictionary<Object,Object>` |    4 |         302 |        312 |    -3% |
| `ReadOnlyDictionary<Object,Object>` |    5 |         357 |        358 |     0% |
| `ReadOnlyDictionary<Object,Object>` |   10 |         631 |        640 |    -1% |
| `ReadOnlyDictionary<Object,Object>` |   25 |        1412 |       1372 |     3% |
| `ReadOnlyDictionary<Object,Object>` |   50 |        2753 |       2642 |     4% |
| `ReadOnlyDictionary<Object,Object>` |  100 |        5280 |       5033 |     5% |
| `ReadOnlyDictionary<String,String>` |    0 |          85 |         92 |    -7% |
| `ReadOnlyDictionary<String,String>` |    1 |         176 |        182 |    -3% |
| `ReadOnlyDictionary<String,String>` |    2 |         238 |        232 |     3% |
| `ReadOnlyDictionary<String,String>` |    3 |         295 |        291 |     1% |
| `ReadOnlyDictionary<String,String>` |    4 |         401 |        415 |    -3% |
| `ReadOnlyDictionary<String,String>` |    5 |         445 |        444 |     0% |
| `ReadOnlyDictionary<String,String>` |   10 |         802 |        772 |     4% |
| `ReadOnlyDictionary<String,String>` |   25 |        1761 |       1708 |     3% |
| `ReadOnlyDictionary<String,String>` |   50 |        3532 |       3414 |     3% |
| `ReadOnlyDictionary<String,String>` |  100 |        6589 |       6723 |    -2% |
| `ReadOnlyDictionary<Int64,Int64>`   |    0 |          73 |         79 |    -8% |
| `ReadOnlyDictionary<Int64,Int64>`   |    1 |         140 |        134 |     5% |
| `ReadOnlyDictionary<Int64,Int64>`   |    2 |         161 |        169 |    -5% |
| `ReadOnlyDictionary<Int64,Int64>`   |    3 |         181 |        195 |    -7% |
| `ReadOnlyDictionary<Int64,Int64>`   |    4 |         223 |        241 |    -7% |
| `ReadOnlyDictionary<Int64,Int64>`   |    5 |         255 |        264 |    -4% |
| `ReadOnlyDictionary<Int64,Int64>`   |   10 |         410 |        404 |     2% |
| `ReadOnlyDictionary<Int64,Int64>`   |   25 |         867 |        896 |    -3% |
| `ReadOnlyDictionary<Int64,Int64>`   |   50 |        1650 |       1660 |    -1% |
| `ReadOnlyDictionary<Int64,Int64>`   |  100 |        3162 |       3161 |     0% |
| `ReadOnlyDictionary<Int32,Int32>`   |    0 |          73 |         73 |     0% |
| `ReadOnlyDictionary<Int32,Int32>`   |    1 |         129 |        128 |     1% |
| `ReadOnlyDictionary<Int32,Int32>`   |    2 |         154 |        161 |    -4% |
| `ReadOnlyDictionary<Int32,Int32>`   |    3 |         174 |        186 |    -6% |
| `ReadOnlyDictionary<Int32,Int32>`   |    4 |         206 |        212 |    -3% |
| `ReadOnlyDictionary<Int32,Int32>`   |    5 |         236 |        232 |     2% |
| `ReadOnlyDictionary<Int32,Int32>`   |   10 |         383 |        394 |    -3% |
| `ReadOnlyDictionary<Int32,Int32>`   |   25 |         806 |        840 |    -4% |
| `ReadOnlyDictionary<Int32,Int32>`   |   50 |        1538 |       1571 |    -2% |
| `ReadOnlyDictionary<Int32,Int32>`   |  100 |        2960 |       3033 |    -2% |